Added a version constant to constants.py.
Added some code to size() so that small canvases are surrounded by a 
margin so as to get around a problem with small windows in MS Windows 
(is this why Processing does the same thing? I wonder...)

Ellipses now use a smaller sampling in order to help speed up some 
scripts.



git-svn-id: http://pyprocessing.googlecode.com/svn/trunk@45 d37874ac-98ab-11de-ac5d-73b9d092dc0a